requirements.txt
python-nmap==0.7.1
google-generativeai==0.3.1
python-dotenv==1.0.0
